fix: use vanilla double precision for Slow Falling distance

+++ src/main/java/io/github/r3neer/alchemicalleather/effect/SlowFallingWear.java
@@ -2,7 +2,7 @@ package io.github.r3neer.alchemicalleather.effect;
 
 /** Small causal helpers for Slow Falling mechanics that are separate from its gravity clamp. */
 public final class SlowFallingWear {
-    private static final float EPS=1.0E-6F;
+    private static final double EPS=1.0E-6D;
     private SlowFallingWear(){}
 
     /**
@@ -10,7 +10,7 @@ public final class SlowFallingWear {
      * Falling only when there was accumulated fall distance to erase, it was actually erased, and
      * Levitation would not have caused the same reset without it.
      */
-    public static boolean fallDistanceResetNeeded(float before,float after,boolean levitation){
-        return Float.isFinite(before)&&Float.isFinite(after)&&before>EPS&&after<=EPS&&!levitation;
+    public static boolean fallDistanceResetNeeded(double before,double after,boolean levitation){
+        return Double.isFinite(before)&&Double.isFinite(after)&&before>EPS&&after<=EPS&&!levitation;
     }
 }
